back/requirements.txt
Flask
Flask-Cors
requests
python-dotenv
PyJWT
passlib
psycopg2-binary
pandas
openpyxl
Werkzeug